requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
cors==1.0.1
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==25.5.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.5
requests-file==2.1.0
SQLAlchemy==2.0.43
tldextract==5.3.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
zope.event==5.1.1
zope.interface==7.2
